# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rest Api documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://rest-api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "getUsersList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserListResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "BaseList": {
                "title": "Base list",
                "description": "Base list model",
                "properties": {
                    "currentPage": {
                        "title": "current_page",
                        "description": "Number of the current page",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "firstPageUrl": {
                        "title": "first_page_url",
                        "description": "Url of first page",
                        "type": "string",
                        "example": "http://rest-api/api/route?page=1"
                    },
                    "from": {
                        "title": "from",
                        "description": "Start id of item on current page",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "lastPage": {
                        "title": "last_page",
                        "description": "Last number page of pagination",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "lastPageUrl": {
                        "title": "last_page_url",
                        "description": "Url of the last page",
                        "type": "string",
                        "example": "http://rest-api/api/route?page=1"
                    },
                    "nextPageUrl": {
                        "title": "next_page_url",
                        "description": "Url of the next page",
                        "type": "string",
                        "example": null
                    },
                    "path": {
                        "title": "path",
                        "description": "Current path",
                        "type": "string",
                        "example": "http://rest-api/api/route"
                    },
                    "perPage": {
                        "title": "per_page",
                        "description": "Count of items per page",
                        "type": "integer",
                        "format": "int64",
                        "example": 10
                    },
                    "prevPageUrl": {
                        "title": "prev_page_url",
                        "description": "Url of the previous page",
                        "type": "string",
                        "example": null
                    },
                    "to": {
                        "title": "to",
                        "description": "Last id of item on current page",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "total": {
                        "title": "total",
                        "description": "Count of items on current page",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Base list"
                }
            },
            "UsersList": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseList"
                    },
                    {
                        "title": "Business list",
                        "description": "Business list model",
                        "properties": {
                            "data": {
                                "title": "Data",
                                "description": "Data wrapper",
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "id": {
                                            "type": "int64",
                                            "example": 2
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "User"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "test@test.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "Business list"
                        }
                    }
                ]
            },
...
            "ErrorResource": {
                "title": "ErrorResource",
                "description": "Error resource",
                "properties": {
                    "status": {
                        "title": "Status",
                        "description": "Status wrapper",
                        "example": "error"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ErrorResource"
                }
            },
            "SuccessResource": {
                "title": "SuccessResource",
                "description": "Success resource",
                "properties": {
                    "status": {
                        "title": "Status",
                        "description": "Status wrapper",
                        "example": "success"
                    },
                    "message": {
                        "title": "Message",
                        "description": "Message wrapper",
                        "example": "The operation was successful."
                    }
                },
                "type": "object",
                "xml": {
                    "name": "SuccessResource"
                }
            },
...
            "ForbiddenResource": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ErrorResource"
                    },
                    {
                        "title": "ForbiddenResource",
                        "description": "Forbidden resource",
                        "properties": {
                            "message": {
                                "title": "Message",
                                "description": "Message wrapper",
                                "example": "Forbidden"
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "ForbiddenResource"
                        }
                    }
                ]
            },
...
            "UnauthenticatedResource": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ErrorResource"
                    },
                    {
                        "title": "UnauthenticatedResource",
                        "description": "Unauthenticated resource",
                        "properties": {
                            "status": {
                                "title": "Status",
                                "description": "Status wrapper",
                                "example": "failed"
                            },
                            "message": {
                                "title": "Message",
                                "description": "Message wrapper",
                                "example": "Unauthenticated"
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "UnauthenticatedResource"
                        }
                    }
                ]
            },
            "UserListResource": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/SuccessResource"
                    },
                    {
                        "title": "UserListResource",
                        "description": "User list resource",
                        "properties": {
                            "data": {
                                "title": "Data",
                                "description": "Data wrapper",
                                "properties": {
                                    "usersList": {
                                        "title": "User list",
                                        "description": "User list wrapper",
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/UsersList"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "UserListResource"
                        }
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://rest-api/oauth/authorize",
                        "tokenUrl": "http://rest-api/oauth/token",
                        "refreshUrl": "http://rest-api/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}